refactor: separate RoL world format parsers and emitters
