+++ CMakeLists.txt
@@ -1002,6 +1002,15 @@ set(WORLD_TOOL_SOURCES
     scripts/world/wtool_lib/rol_state_periodic_profiles.py
     scripts/world/wtool_lib/rol_soc.py
     scripts/world/wtool_lib/rol_special.py
+    scripts/world/wtool_lib/rol_conversion_types.py
+    scripts/world/wtool_lib/rol_source_common.py
+    scripts/world/wtool_lib/rol_transform_common.py
+    scripts/world/wtool_lib/rol_mobiles.py
+    scripts/world/wtool_lib/rol_objects.py
+    scripts/world/wtool_lib/rol_rooms.py
+    scripts/world/wtool_lib/rol_zones.py
+    scripts/world/wtool_lib/rol_shops.py
+    scripts/world/wtool_lib/rol_quests.py
     scripts/world/wtool_lib/rol_source.py
     scripts/world/wtool_lib/rol_transform.py
     scripts/world/wtool_lib/rooms.py
