+++ scripts/world/wtool_lib/rol_phase8.py
@@ -49,6 +49,39 @@ _CODE_EVIDENCE_PATHS = (
     "scripts/world/wtool_lib/rol_persistence.py",
     "scripts/world/wtool_lib/rol_persistence_check.py",
     "scripts/world/wtool_lib/rol_transform.py",
+    "scripts/world/wtool_lib/rol_source.py",
+    "scripts/world/wtool_lib/rol_conversion_types.py",
+    "scripts/world/wtool_lib/rol_source_common.py",
+    "scripts/world/wtool_lib/rol_transform_common.py",
+    "scripts/world/wtool_lib/rol_mobiles.py",
+    "scripts/world/wtool_lib/rol_objects.py",
+    "scripts/world/wtool_lib/rol_rooms.py",
+    "scripts/world/wtool_lib/rol_zones.py",
+    "scripts/world/wtool_lib/rol_shops.py",
+    "scripts/world/wtool_lib/rol_quests.py",
+    "scripts/world/wtool_lib/rol_soc.py",
+    "scripts/world/wtool_lib/rol_mobile_identity.py",
+    "scripts/world/wtool_lib/rol_mob_calculator.py",
+    "scripts/world/wtool_lib/rol_weapon_mapping.py",
+    "scripts/world/wtool_lib/rol_weapon_table.py",
+    "scripts/world/wtool_lib/rol_weapon_overrides.json",
+    "scripts/world/wtool_lib/rol_special.py",
+    "scripts/world/wtool_lib/rol_periodic_profiles.py",
+    "scripts/world/wtool_lib/rol_state_periodic_profiles.py",
+    "scripts/world/wtool_lib/rol_pilot_build.py",
+    "scripts/world/wtool_lib/rol_identity.py",
+    "scripts/world/wtool_lib/rol_inventory.py",
+    "scripts/world/wtool_lib/rol_discovery.py",
+    "scripts/world/wtool_lib/source.py",
+    "scripts/world/wtool_lib/flags.py",
+    "scripts/world/wtool_lib/constants.py",
+    "scripts/world/rol_conversion_policy.json",
+    "scripts/world/rol_source_race_registry.json",
+    "scripts/world/wtool_constants.json",
+    "bin/rol_mob_calculator",
+    "util/rol_mob_calculator.c",
+    "src/structs.h",
+    "src/combat/assign_wpn_armor.c",
     "src/db.c",
     "src/spec/spec_assign_mobiles.c",
     "src/spec/spec_assign_objects.c",
